Split out generic palm_utils.py for use across all PALM variants
palm_utils provides the common code for Solcast and historical data gathering and SoC forecasting. A smaller wrapper is then used for each PALM variant.

This code is the Beta version of PALM v1.1.0
